+++ frontend/src/i18n/index.js
@@ -1,0 +1,4 @@
+
+export { default as pl } from './pl.js';
+export { default as en } from './en.js';
+export { default as de } from './de.js';